Chore/CI Actions (#30)
* repo: add black and ruff github actions

Added ruff and black github actions, they are separate as one is
independent from the other (both can show issues, so both need to be ran
even if the other has failed), and the yml file describing each one is
very simple and standard.

* format: format the current codebase to standard black formmater

Used black to format the file kobra.py.

* fix: black action call to stable version

* code: join CI jobs in single workflow

* fix: only run ci on pr

Only run ci on PR, as main branch should be protected, and it is only a
validation github action

* format after fixing conflicts

* docs: documenting development requirements

uv is optional as all packages can be installed system wise but ruff and
black are needed for pull request merging.

CONTRIBUTING and conventions were also updated to describe this need