feat(packaging): set proper entrypoint and __main__.py (#337)
As it is right now, the project is not a valid python package. The main file being outside `src` is out of standard, and it was missing a proper entrypoint.

This PR:

1. Moves the `main` function from `cobra.py` to `src/core/__init__.py`
2. Adds a proper `__main__.py` that calls it
3. Fixes imports, which should be relative to `src` according to standard
4. Adds an entrypoint, which allows for `pip install .` and `uv run naja` (and corresponding updated docs)
5. Correctly marks the project as installable

Some implications: users can now use standard tooling with the project. For example, `pipx install https://github.com/fossguild/naja` can install the program directly.

Co-authored-by: Danielle Modesti <dani.modesti@usp.br>